requirements.txt
numpy
pyeda
pyyaml
networkx